feat(infra): split dashboard api.ts into per-resource modules (#261 Move 2)
Breaks the 1095-line monolithic api.ts into 12 focused files under
apps/dashboard/src/lib/api/: client.ts, types.ts, sessions.ts,
credentials.ts, pats.ts, workspaces.ts, governance.ts, spine.ts,
artifacts.ts, workflows.ts, issues.ts, pulls.ts, registry.ts, and
index.ts. The original api.ts becomes a one-line re-export so all
existing `import { ... } from '@/lib/api'` calls remain unchanged.
