+++ apps/dashboard/src/lib/api/artifacts.ts
@@ -1,0 +1,33 @@
+import { request } from './client';
+import type {
+  SpineArtifact,
+  ArtifactDetail,
+  RegisterArtifactRequest,
+  ArtifactActionRequest,
+  ArtifactActionResponse,
+  OverrideGateRequestBody,
+} from './types';
+
+export const artifacts = {
+  getArtifact: (id: string) => request<{ artifact: ArtifactDetail }>(`/spine/artifacts/${id}`),
+  registerArtifact: (req: RegisterArtifactRequest) =>
+    request<{ artifact: SpineArtifact }>('/spine/artifacts', {
+      method: 'POST',
+      body: JSON.stringify(req),
+    }),
+  retryArtifact: (id: string, body: ArtifactActionRequest = {}) =>
+    request<ArtifactActionResponse>(`/spine/artifacts/${id}/retry`, {
+      method: 'POST',
+      body: JSON.stringify(body),
+    }),
+  abortArtifact: (id: string, body: ArtifactActionRequest = {}) =>
+    request<ArtifactActionResponse>(`/spine/artifacts/${id}/abort`, {
+      method: 'POST',
+      body: JSON.stringify(body),
+    }),
+  overrideGate: (id: string, body: OverrideGateRequestBody = {}) =>
+    request<ArtifactActionResponse>(`/spine/artifacts/${id}/override-gate`, {
+      method: 'POST',
+      body: JSON.stringify(body),
+    }),
+};
